+++ src/chains/Scroll/MetaTxn.sol
@@ -36,13 +36,13 @@ contract ScrollSettlerMetaTxn is SettlerMetaTxn, ScrollMixin {
     }
 
     // Solidity inheritance is stupid
-    function _dispatch(uint256 i, uint256 action, bytes calldata data)
+    function _dispatch(uint256 i, uint256 action, bytes calldata data, AllowedSlippage memory slippage)
         internal
         virtual
-        override(SettlerAbstract, SettlerBase, ScrollMixin)
+        override(SettlerBase, ScrollMixin)
         returns (bool)
     {
-        return super._dispatch(i, action, data);
+        return super._dispatch(i, action, data, slippage);
     }
 
     function _isRestrictedTarget(address target)