+++ src/core/DodoV1.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.25;
 
 import {IERC20} from "@forge-std/interfaces/IERC20.sol";
-import {SettlerAbstract} from "../SettlerAbstract.sol";
+import {SettlerSwapAbstract} from "../SettlerAbstract.sol";
 import {revertTooMuchSlippage} from "./SettlerErrors.sol";
 import {SafeTransferLib} from "../vendor/SafeTransferLib.sol";
 import {UnsafeMath} from "../utils/UnsafeMath.sol";
@@ -344,7 +344,7 @@ abstract contract DodoSellHelper {
     }
 }
 
-abstract contract DodoV1 is SettlerAbstract, DodoSellHelper {
+abstract contract DodoV1 is SettlerSwapAbstract, DodoSellHelper {
     using UnsafeMath for uint256;
     using SafeTransferLib for IERC20;
     using FastDodoV1 for IDodoV1;