+++ src/core/UniswapV3Fork.sol
@@ -8,7 +8,7 @@ import {UnsafeMath} from "../utils/UnsafeMath.sol";
 import {Panic} from "../utils/Panic.sol";
 import {SafeTransferLib} from "../vendor/SafeTransferLib.sol";
 import {AddressDerivation} from "../utils/AddressDerivation.sol";
-import {SettlerAbstract} from "../SettlerAbstract.sol";
+import {SettlerSwapAbstract} from "../SettlerAbstract.sol";
 
 import {revertTooMuchSlippage} from "./SettlerErrors.sol";
 
@@ -33,7 +33,7 @@ interface IUniswapV3Pool {
     ) external returns (int256 amount0, int256 amount1);
 }
 
-abstract contract UniswapV3Fork is SettlerAbstract {
+abstract contract UniswapV3Fork is SettlerSwapAbstract {
     using Ternary for bool;
     using UnsafeMath for uint256;
     using UnsafeMath for int256;