+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Add Thena UniV3 Algebra style fork to Bnb with fork ID 37
 * EkuboV2 VIP actions `EKUBO_VIP` and `METATXN_EKUBO_VIP` are not
   supported anymore and were removed from `ISettlerActions.sol`.
+* Add new `CHECK_SLIPPAGE` action to taker-submitted Settlers
 * Update the following chains to the Osaka EVM hardfork:
   * Arbitrum
   * Polygon